Automatically cancel payment pending orders after selected number of days (#59)
### Added

- Option to automatically cancel payment pending orders after selected number of days

### Changed

- Uses Tpay OpenApi SDK 2.0 

### Fixed

- Resolves issue with translations loading to early (produced warning to logs)

### Removed

- Dependency on additional constants added to `wp-config.php` file added during plugin initialization

+++ includes/Helpers/CardHelper.php
@@ -8,8 +8,12 @@ class CardHelper
 
     public function save_card(array $card): bool
     {
-        $hash = sha1($card['card_hash'].WP_TPAY_HASH);
-        DatabaseConnection::delete(self::TABLE, ['hash' => $hash, 'source_order' => null]);
+        if (defined('WP_TPAY_HASH')) {
+            $hash = sha1($card['card_hash'].WP_TPAY_HASH);
+            DatabaseConnection::delete(self::TABLE, ['hash' => $hash, 'source_order' => null]);
+        } else {
+            $hash = md5($card['card_hash']);
+        }
         $card_id = DatabaseConnection::queryVar('SELECT id FROM %i WHERE hash = %s', self::TABLE, $hash);
 
         if (!$card_id) {
